BUILD : 1.7.6 (499) BUGS : NOTES : Fixed several compiler warnings with make strict, removed 2 deprecated config vars.
git-svn-id: svn://svn.anope.org/anope/trunk@499 31f1291d-b8d6-0310-a050-a5561fc1590b


git-svn-id: http://anope.svn.sourceforge.net/svnroot/anope/trunk@353 5417fbe8-f217-4b02-8779-1006273d7864

+++ src/unreal32.c
@@ -1506,7 +1506,7 @@ int anope_event_netinfo(char *source, int ac, char **av)
 void anope_cmd_netinfo(int ac, char **av)
 {
     send_cmd(NULL, "%s %ld %ld %d %s 0 0 0 :%s",
-             send_token("NETINFO", "AO"), maxusercnt,
+             send_token("NETINFO", "AO"), (long int) maxusercnt,
              (long int) time(NULL), atoi(av[2]), av[3], av[7]);
 }
